Display product in Card view using for each loop
Made changes to the Card view to display all the 6 products by java tags in the html using for each loop.
Started on the layout /look and feel for the cart page using a table and font Awesome and bootstrap. Added a some script in the header if required and also style inline for the page.